+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    - map
    - filter
    - foldl
+   - mempty and mappend
 
 ## [0.5.0] - 2020-11-10
 ### Added